Move Process.launch_next_process to Service.launch_next_process
